requirements.txt
fastmcp
requests
pydub
